src/sensor-simulator/requirements.txt
paho-mqtt>=2.1.0
numpy>=1.26.0
